DEV: Refactor subclasses in `ThemeSettingsManager` to individual files
Why this change?

One Ruby class per file improves readability
