+++ lib/theme_settings_manager/list.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class ThemeSettingsManager::List < ThemeSettingsManager
+  def list_type
+    @opts[:list_type]
+  end
+end
