+++ spec/fixtures/theme_settings/valid_settings.yaml
@@ -84,16 +84,3 @@ valid_json_schema_setting:
 causes_refresh:
   default: ""
   refresh: true
-
-valid_objects_setting:
-  type: objects
-  default:
-    - title: "Some title"
-      description: "Some description"
-  schema:
-    name: "Some Object"
-    fields:
-      title:
-        type: string
-      description:
-        type: string
