+++ spec/lib/theme_settings_manager_spec.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-require "theme_settings_manager"
-
 RSpec.describe ThemeSettingsManager do
   let!(:theme) { Fabricate(:theme) }
 
@@ -11,8 +9,6 @@ RSpec.describe ThemeSettingsManager do
     theme.save!
     theme.settings
   end
-
-  before { SiteSetting.experimental_objects_type_for_theme_settings = true }
 
   describe "Enum" do
     it "only accepts values from its choices" do
@@ -184,23 +180,6 @@ RSpec.describe ThemeSettingsManager do
           expect(upload_setting.value).to eq(Discourse.store.cdn_url(upload.url))
         end
       end
-    end
-  end
-
-  describe ThemeSettingsManager::Objects do
-    it "can store a list of objects" do
-      objects_setting = theme_settings[:valid_objects_setting]
-
-      expect(objects_setting.value).to eq(
-        [{ "title" => "Some title", "description" => "Some description" }],
-      )
-
-      objects_setting.value = [{ title: "title 1", description: "description 1" }]
-      objects_setting = theme.reload.settings[:valid_objects_setting]
-
-      expect(objects_setting.value).to eq(
-        [{ "title" => "title 1", "description" => "description 1" }],
-      )
     end
   end
 end
